Add AudioPlaylistEntry Component
