+++ client/src/components/Home/MediaTypes/Music.js
@@ -1,33 +1,3 @@
-// import React, { Component } from 'react';
-// import MusicPlayer from './MusicPlayer';
-
-// export const TRACKS = [
-//   {
-//     title: 'Stressed Out',
-//     artist: 'Twenty One Pilots',
-//     albumArtUrl: "http://36.media.tumblr.com/14e9a12cd4dca7a3c3c4fe178b607d27/tumblr_nlott6SmIh1ta3rfmo1_1280.jpg",
-//     audioUrl: "http://dl.fazmusics.in/Ali/music/aban/hot%20100%20.7%20nov%202015(128)/Twenty%20One%20Pilots%20-%20Stressed%20Out.mp3",
-//   },
-//   {
-//     title: 'Love Yourself',
-//     artist: 'Justin Bieber',
-//     albumArtUrl: "http://arrestedmotion.com/wp-content/uploads/2015/10/JB_Purpose-digital-deluxe-album-cover_lr.jpg",
-//     audioUrl: 'http://srv2.dnupload.com/Music/Album/Justin%20Bieber%20-%20Purpose%20(Deluxe%20Version)%20(320)/Justin%20Bieber%20-%20Purpose%20(Deluxe%20Version)%20128/05%20Love%20Yourself.mp3',
-//   },
-//   {
-//     title: 'Hotline Bling',
-//     artist: 'Drake',
-//     albumArtUrl: 'https://upload.wikimedia.org/wikipedia/commons/c/c9/Drake_-_Hotline_Bling.png',
-//     audioUrl: 'http://dl2.shirazsong.org/dl/music/94-10/CD%201%20-%20Best%20of%202015%20-%20Top%20Downloads/03.%20Drake%20-%20Hotline%20Bling%20.mp3',
-//   },
-// ];
-
-// export default class Music extends Component {
-//   render() {
-//     return <MusicPlayer tracks={TRACKS} />
-//   }
-// }
-
 import React, { Component } from 'react';
 import {
 	Dimensions,
@@ -38,6 +8,8 @@ import {
 	TouchableHighlight,
 	View,
 } from 'react-native';
+import AudioPlaylistEntry from './AudioPlaylistEntry';
+
 import Slider from 'react-native-slider';
 import { Asset, Audio, Font } from 'expo';
 import { MaterialIcons, Ionicons, Feather } from '@expo/vector-icons';
@@ -65,411 +37,76 @@ const PLAYLIST = [
 		'Podington Bear - “Rubber Robot”',
 		'https://s3.amazonaws.com/exp-us-standard/audio/playlist-example/Podington_Bear_-_Rubber_Robot.mp3',
 		'https://cdn-img.jamendo.com/albums/s166/166372/covers/1.300.jpg?du=1529657778'
+  ),
+  new PlaylistItem(
+		'Comfort Fit - “Sorry”',
+		'https://s3.amazonaws.com/exp-us-standard/audio/playlist-example/Comfort_Fit_-_03_-_Sorry.mp3',
+		'https://images.shazam.com/coverart/t55537491-b955650891_s400.jpg'
+	),
+	new PlaylistItem(
+		'Mildred Bailey – “All Of Me”',
+		'https://ia800304.us.archive.org/34/items/PaulWhitemanwithMildredBailey/PaulWhitemanwithMildredBailey-AllofMe.mp3',
+		'https://tinyurl.com/y9u8e437'
+	),
+	new PlaylistItem(
+		'Podington Bear - “Rubber Robot”',
+		'https://s3.amazonaws.com/exp-us-standard/audio/playlist-example/Podington_Bear_-_Rubber_Robot.mp3',
+		'https://cdn-img.jamendo.com/albums/s166/166372/covers/1.300.jpg?du=1529657778'
+  ),
+  new PlaylistItem(
+		'Comfort Fit - “Sorry”',
+		'https://s3.amazonaws.com/exp-us-standard/audio/playlist-example/Comfort_Fit_-_03_-_Sorry.mp3',
+		'https://images.shazam.com/coverart/t55537491-b955650891_s400.jpg'
+	),
+	new PlaylistItem(
+		'Mildred Bailey – “All Of Me”',
+		'https://ia800304.us.archive.org/34/items/PaulWhitemanwithMildredBailey/PaulWhitemanwithMildredBailey-AllofMe.mp3',
+		'https://tinyurl.com/y9u8e437'
+	),
+	new PlaylistItem(
+		'Podington Bear - “Rubber Robot”',
+		'https://s3.amazonaws.com/exp-us-standard/audio/playlist-example/Podington_Bear_-_Rubber_Robot.mp3',
+		'https://cdn-img.jamendo.com/albums/s166/166372/covers/1.300.jpg?du=1529657778'
 	),
 ];
 
 const { width: DEVICE_WIDTH, height: DEVICE_HEIGHT } = Dimensions.get('window');
-// const BACKGROUND_COLOR = 'rgb(26,201,141)';
-const DISABLED_OPACITY = 0.5;
-const FONT_SIZE = 14;
-const LOADING_STRING = 'Loading...';
-const BUFFERING_STRING = 'Buffering...';
-const RATE_SCALE = 3.0;
 
 export default class Music extends Component {
 	constructor(props) {
 		super(props);
 		this.index = 0;
-		this.isSeeking = false;
-		this.shouldPlayAtEndOfSeek = false;
 		this.playbackInstance = undefined;
 		this.state = {
-			playbackInstanceName: LOADING_STRING,
-			playbackInstancePosition: undefined,
-			playbackInstanceDuration: undefined,
-			shouldPlay: false,
-			isPlaying: false,
-			isBuffering: false,
-			isLoading: true,
-			fontLoaded: false,
-			volume: 1.0,
-			rate: 1.0,
-			portrait: undefined,
+      playlist: PLAYLIST
+			// playbackInstanceName: LOADING_STRING,
+			// playbackInstancePosition: undefined,
+			// playbackInstanceDuration: undefined,
+			// shouldPlay: false,
+			// isPlaying: false,
+			// isBuffering: false,
+			// isLoading: true,
+			// fontLoaded: false,
+			// volume: 1.0,
+			// rate: 1.0,
+			// portrait: undefined,
 		};
-	}
-
-	componentDidMount() {
-		Audio.setAudioModeAsync({
-			allowsRecordingIOS: false,
-			interruptionModeIOS: Audio.INTERRUPTION_MODE_IOS_DO_NOT_MIX,
-			playsInSilentModeIOS: true,
-			shouldDuckAndroid: true,
-			interruptionModeAndroid: Audio.INTERRUPTION_MODE_ANDROID_DO_NOT_MIX,
-		});
-		(async () => {
-			await Font.loadAsync({
-				roboto: require('../../../../assets/fonts/Roboto.ttf'),
-			});
-			this.setState({ fontLoaded: true });
-		})();
-
-		this._loadNewPlaybackInstance(false);
-	}
-
-	async _loadNewPlaybackInstance(playing) {
-		if (this.playbackInstance != null) {
-			await this.playbackInstance.unloadAsync();
-			this.playbackInstance.setOnPlaybackStatusUpdate(null);
-			this.playbackInstance = null;
-		}
-
-		const source = { uri: PLAYLIST[this.index].uri };
-		const initialStatus = {
-			shouldPlay: playing,
-			rate: this.state.rate,
-			volume: this.state.volume,
-		};
-
-		const { sound, status } = await Audio.Sound.create(
-			source,
-			initialStatus,
-			this._onPlaybackStatusUpdate
-		);
-		this.playbackInstance = sound;
-
-		this._updateScreenForLoading(false);
-	}
-
-	_updateScreenForLoading(isLoading) {
-		if (isLoading) {
-			this.setState({
-				isPlaying: false,
-				playbackInstanceName: LOADING_STRING,
-				playbackInstanceDuration: null,
-				playbackInstancePosition: null,
-				isLoading: true,
-			});
-		} else {
-			this.setState({
-				playbackInstanceName: PLAYLIST[this.index].name,
-				portrait: PLAYLIST[this.index].image,
-				isLoading: false,
-			});
-		}
-	}
-
-	_onPlaybackStatusUpdate = status => {
-		if (status.isLoaded) {
-			this.setState({
-				playbackInstancePosition: status.positionMillis,
-				playbackInstanceDuration: status.durationMillis,
-				shouldPlay: status.shouldPlay,
-				isPlaying: status.isPlaying,
-				isBuffering: status.isBuffering,
-				rate: status.rate,
-				volume: status.volume,
-			});
-			if (status.didJustFinish) {
-				this._advanceIndex(true); 
-				this._updatePlaybackInstanceForIndex(true);
-			}
-		} else {
-			if (status.error) {
-				console.log(`FATAL PLAYER ERROR: ${status.error}`);
-			}
-		}
-	};
-
-	_advanceIndex(forward) {
-		this.index =
-			(this.index + (forward ? 1 : PLAYLIST.length - 1)) %
-			PLAYLIST.length;
-	}
-
-	async _updatePlaybackInstanceForIndex(playing) {
-		this._updateScreenForLoading(true);
-
-		this._loadNewPlaybackInstance(playing);
-	}
-
-	_onPlayPausePressed = () => {
-		if (this.playbackInstance != null) {
-			if (this.state.isPlaying) {
-				this.playbackInstance.pauseAsync();
-			} else {
-				this.playbackInstance.playAsync();
-			}
-		}
-	};
-
-	_onStopPressed = () => {
-		if (this.playbackInstance != null) {
-			this.playbackInstance.stopAsync();
-		}
-	};
-
-	_onForwardPressed = () => {
-		if (this.playbackInstance != null) {
-			this._advanceIndex(true);
-			this._updatePlaybackInstanceForIndex(this.state.shouldPlay);
-		}
-	};
-
-	_onBackPressed = () => {
-		if (this.playbackInstance != null) {
-			this._advanceIndex(false);
-			this._updatePlaybackInstanceForIndex(this.state.shouldPlay);
-		}
-	};
-
-	_onVolumeSliderValueChange = value => {
-		if (this.playbackInstance != null) {
-			this.playbackInstance.setVolumeAsync(value);
-		}
-	};
-
-	_trySetRate = async rate => {
-		if (this.playbackInstance != null) {
-			try {
-				await this.playbackInstance.setRateAsync(rate);
-			} catch (error) {
-				// Rate changing could not be performed, possibly because the client's Android API is too old.
-			}
-		}
-	};
-
-	_onRateSliderSlidingComplete = async value => {
-		this._trySetRate(value * RATE_SCALE);
-	};
-
-	_onSeekSliderValueChange = value => {
-		if (this.playbackInstance != null && !this.isSeeking) {
-			this.isSeeking = true;
-			this.shouldPlayAtEndOfSeek = this.state.shouldPlay;
-			this.playbackInstance.pauseAsync();
-		}
-	};
-
-	_onSeekSliderSlidingComplete = async value => {
-		if (this.playbackInstance != null) {
-			this.isSeeking = false;
-			const seekPosition = value * this.state.playbackInstanceDuration;
-			if (this.shouldPlayAtEndOfSeek) {
-				this.playbackInstance.playFromPositionAsync(seekPosition);
-			} else {
-				this.playbackInstance.setPositionAsync(seekPosition);
-			}
-		}
-	};
-
-	_getSeekSliderPosition() {
-		if (
-			this.playbackInstance != null &&
-			this.state.playbackInstancePosition != null &&
-			this.state.playbackInstanceDuration != null
-		) {
-			return (
-				this.state.playbackInstancePosition /
-				this.state.playbackInstanceDuration
-			);
-		}
-		return 0;
-	}
-
-	_getMMSSFromMillis(millis) {
-		const totalSeconds = millis / 1000;
-		const seconds = Math.floor(totalSeconds % 60);
-		const minutes = Math.floor(totalSeconds / 60);
-
-		const padWithZero = number => {
-			const string = number.toString();
-			if (number < 10) {
-				return '0' + string;
-			}
-			return string;
-		};
-		return padWithZero(minutes) + ':' + padWithZero(seconds);
-	}
-
-	_getTimestamp() {
-		if (
-			this.playbackInstance != null &&
-			this.state.playbackInstancePosition != null &&
-			this.state.playbackInstanceDuration != null
-		) {
-			return `${this._getMMSSFromMillis(
-				this.state.playbackInstancePosition
-			)} / ${this._getMMSSFromMillis(
-				this.state.playbackInstanceDuration
-			)}`;
-		}
-		return '';
 	}
 
 	render() {
-		return !this.state.fontLoaded ? (
+		return false ? (
 			<View />
 		) : (
       <ImageBackground source={require('../../../../assets/img/gradient-background-image.png')} style={{ width: '100%', height: '100%' }}>
+        <View style={styles.container}>
 
-			<View style={styles.container}>
-				<View style={styles.portraitContainer}>
-					<Image
-						style={styles.portrait}
-						source={{
-							uri: this.state.portrait,
-						}}
-					/>
-				</View>
-				<View style={styles.detailsContainer}>
-					<Text> 
-						{this.state.playbackInstanceName}
-					</Text>
-					<Text>
-						{this.state.isBuffering ? (
-							BUFFERING_STRING
-						) : (
-							this._getTimestamp()
-						)}
-					</Text>
-				</View>
-				<View
-					style={[
-						styles.buttonsContainerBase,
-						styles.buttonsContainerTopRow,
-						{
-							opacity: this.state.isLoading
-								? DISABLED_OPACITY
-								: 1.0,
-						},
-					]}
-				>
-					<TouchableHighlight
-						// underlayColor={BACKGROUND_COLOR}
-						style={styles.wrapper}
-						onPress={this._onBackPressed}
-						disabled={this.state.isLoading}
-					>
-						<View>
-							<Feather
-								name="skip-back"
-								size={40}
-								color="#ffffff"
-							/>
-						</View>
-					</TouchableHighlight>
-					<TouchableHighlight
-						// underlayColor={BACKGROUND_COLOR}
-						style={styles.wrapper}
-						onPress={this._onPlayPausePressed}
-						disabled={this.state.isLoading}
-					>
-						<View>
-							{this.state.isPlaying ? (
-								<Feather
-									name="pause"
-									size={60}
-									color="#ffffff"
-								/>
-							) : (
-								<Feather
-									name="play"
-									size={60}
-									color="#ffffff"
-								/>
-							)}
-						</View>
-					</TouchableHighlight>
-					{/* <TouchableHighlight
-						underlayColor={BACKGROUND_COLOR}
-						style={styles.wrapper}
-						onPress={this._onStopPressed}
-						disabled={this.state.isLoading}
-					>
-						<View> // DON'T THINK I WANT A STOP CIRCLE BUT KEEPING IT COMMENTED IN CASE REMOVING IT BREAKS SOMETHING I DONT' YET UNDERSTAND
-							<Feather
-								name="stop-circle"
-								size={40}
-								color="#ffffff"
-							/>
-						</View>
-					</TouchableHighlight> */}
-					<TouchableHighlight
-						// underlayColor={BACKGROUND_COLOR}
-						style={styles.wrapper}
-						onPress={this._onForwardPressed}
-						disabled={this.state.isLoading}
-					>
-						<View>
-							<Feather
-								name="skip-forward"
-								size={40}
-								color="#ffffff"
-							/>
-						</View>
-					</TouchableHighlight>
-				</View>
-				<View
-					style={[
-						styles.playbackContainer,
-						{
-							opacity: this.state.isLoading
-								? DISABLED_OPACITY
-								: 1.0,
-						},
-					]}
-				>
-					<Slider
-						style={styles.playbackSlider}
-						value={this._getSeekSliderPosition()}
-						onValueChange={this._onSeekSliderValueChange}
-						onSlidingComplete={this._onSeekSliderSlidingComplete}
-						thumbTintColor="#ffffff"
-						minimumTrackTintColor="#ffffff"
-						disabled={this.state.isLoading}
-					/>
-				</View>
-				<View
-					style={[
-						styles.buttonsContainerBase,
-						styles.buttonsContainerMiddleRow,
-					]}
-				>
-					<View style={styles.volumeContainer}>
-						<View>
-							<Ionicons
-                style={styles.volumeDownIcon}
-								name="ios-volume-down-outline"
-								size={40}
-								color="#ffffff"
-							/>
-						</View>
-						<Slider
-							style={styles.volumeSlider}
-							value={1}
-							onValueChange={this._onVolumeSliderValueChange}
-							thumbTintColor="#ffffff"
-							minimumTrackTintColor="#ffffff"
-						/>
-						<View>
-							<Ionicons
-                style={styles.volumeUpIcon}
-								name="ios-volume-up-outline"
-								size={40}
-								color="#ffffff"
-							/>
-						</View>
-					</View>
-				</View>
-				<View
-					style={[
-						styles.buttonsContainerBase,
-						styles.buttonsContainerBottomRow,
-					]}
-				>
-
-				</View>
-			</View>
+          {this.state.playlist.map((track, i) =>
+            <AudioPlaylistEntry
+              key={i + Math.random()} 
+              track={this.state.playlist[i]}
+            />
+          )}
+        </View>
       </ImageBackground>
 		);
 	}
@@ -477,82 +114,11 @@ export default class Music extends Component {
 
 const styles = StyleSheet.create({
 	container: {
-		flex: 1,
-		flexDirection: 'column',
-		justifyContent: 'space-between',
-		alignItems: 'center',
-		alignSelf: 'stretch',
-		// backgroundColor: BACKGROUND_COLOR,
-	},
-	portraitContainer: {
-		marginTop: 30,
-	},
-	portrait: {
-		height: 240,
-		width: 240,
-	},
-	detailsContainer: {
-		height: 40,
-		marginTop: 27,
-    alignItems: 'center',
+    height: DEVICE_HEIGHT,
+    width: DEVICE_WIDTH, 
+		justifyContent: 'center',
 
-	},
-	playbackContainer: {
-		flex: 1,
-		flexDirection: 'column',
-    justifyContent: 'space-between',
-    marginTop: 15,
-		alignItems: 'center',
-		alignSelf: 'stretch',
-	},
-	playbackSlider: {
-    alignSelf: 'stretch',
-		marginLeft: 35,
-    marginRight: 35,
-    marginTop: 10,
-	},
-	text: {
-    color: '#ffffff',
-		fontSize: FONT_SIZE,
-		minHeight: FONT_SIZE,
-	},
-	buttonsContainerBase: {
-		flex: 1,
-		flexDirection: 'row',
-		alignItems: 'center',
-		justifyContent: 'space-between',
-	},
-	buttonsContainerTopRow: {
-		maxHeight: 80,
-		minWidth: DEVICE_WIDTH / 2.0,
-		maxWidth: DEVICE_WIDTH / 2.0,
-	},
-	buttonsContainerMiddleRow: {
-		maxHeight: 40,
-		alignSelf: 'stretch',
-		paddingRight: 20,
-	},
-	volumeContainer: {
-		flex: 1,
-		flexDirection: 'row',
-		alignItems: 'center',
-		justifyContent: 'space-between',
-		minWidth: DEVICE_WIDTH - 40,
-    maxWidth: DEVICE_WIDTH - 40,
-    paddingRight: 25,
-    paddingLeft: 25,
-    marginTop: 10,
-	},
-	volumeSlider: {
-		width: DEVICE_WIDTH - 110,
-  },
-  volumeDownIcon: {
-    paddingRight: 10,
-  },
-  volumeUpIcon: {
-    paddingLeft: 10,
-  },
-	buttonsContainerBottomRow: {
-		alignSelf: 'stretch',
-	},
+		// alignSelf: 'stretch',
+		// backgroundColor: BACKGROUND_COLOR,
+  }
 });
